Added gradle information for jar file
This means we can run the jar file by itself instead of needing to use
gradle run. This does not seem to be a long term solution though, since
the jar appears to be stored in a temporary cache. It'll at least work on
the computer it was compiled on, but probably won't work if the jar file
is moved to another computer.

The jar file is put in the build/libs directory, and is run using java's
-jar command line option:

`java -jar build/libs/connect-four-incubating.jar`

Gradle will append the version to the main file name.