Adding the makefile back
Although gradle is awesome, the csegrid does not have gradle, nor
do we have admin privileges to change that. Since in most cases the
csegrid is used as a standard for grader testing, I will keep it.

The makefile will not work for windows, but gradle will. At the
moment the makefile does not respect the gson dependency, but that
can be changed in the future. Making the driver conform to either
just the csegrid or for all platforms is not required, but I find
this to be an interesting excercise.

We can also just decide to remove this file later.